API Reference

ProtocolDataUnits.PDU — Type

Parent data type for all PDUs.

ProtocolDataUnits.byteorder — Function

byteorder(::Type{T})

Byte order used for PDUs of type T. Defaults to BIG_ENDIAN. To change byte order, define a method for this function.

Example:

ProtocolDataUnits.byteorder(::Type{MyPDU}) = LITTLE_ENDIAN

byteorder(::Type{T}, ::Val{s})

Byte order used for PDUs of type T for field s. Defaults to the same byte order as the PDU. To change byte order, define a method for this function.

Example:

ProtocolDataUnits.byteorder(::Type{MyPDU}, ::Val{:myfield}) = LITTLE_ENDIAN

Base.length — Method

length(::Type{T}, ::Val{s}, info::PDUInfo)

Length of field s in PDU of type T. Defaults to nothing, which indicates that the length is not known, and wire-encoding is used to store length as part of PDU. The length is specified in number of elements for vectors, and number of bytes for strings.

Examples:

# length of field x is 4 bytes less than length of PDU
Base.length(::Type{MyPDU}, ::Val{:x}, info) = info.length - 4

# length of field x is given by the value of field n in the PDU
Base.length(::Type{MyPDU}, ::Val{:x}, info) = info.get(:n)

ProtocolDataUnits.PDUInfo — Type

PDU information with fields:

length: length of PDU in bytes, if known, missing otherwise

get: function that returns value of field s in the PDU

Base.read — Method

read(io::IO, T::PDU)
read(io::IO, T::PDU; nbytes)

Decodes a vector of bytes from stream io to give a PDU. If nbytes is specified, the PDU is assumed to be of length nbytes bytes.

Base.write — Method

write(io::IO, pdu::PDU)

Encodes a PDU into a vector of bytes written to stream io.

Base.Vector — Method

Vector{UInt8}(pdu::PDU)

Encodes a PDU into a vector of bytes.